+++ samples/bluetooth/peripheral/CMakeLists.txt
@@ -6,7 +6,5 @@ project(peripheral)
 
 target_sources(app PRIVATE
   src/main.c
-  ../gatt/cts.c
+  src/cts.c
 )
-
-zephyr_library_include_directories($ENV{ZEPHYR_BASE}/samples/bluetooth)
