+++ samples/bluetooth/peripheral/src/cts.c
