+++ samples/bluetooth/peripheral/src/cts.h
